Check for premium content when using Token shortcuts
Tables for Mutations and the Wrath of the Gods are only available in the premium `wfrp4e-core` content module. If this is not installed and activated when using the related Token Hud Extension shortcuts, a 'Table not found' error would previously be thrown by the system.

This change verifies whether the module is installed, and if not, ignores the shortcut. For now, this is a silent bypass. Going forward this may be enhanced to provide a more specific error message, such as 'The Physical Mutation table is not available. Install and activate the premium core content module to enable this.'

The constant `wfrp4eContent` may be broadened for other premium content checks as needed if the Token Hud Extensions suite is extended. It may be useful to pull it out of the Token Hud Extension and broaden its scope to make available for other module functionality. No pressing need at this time though.