Restrict who can send and be sent a Dark Whisper
Adds an explicit restriction to ensure that only GM users can execute the macro. If the current user is not a GM, an error is shown in the user interface. Returning the error from within a macro needs the code to be restructured into a discrete function.

The target selection list is filtered to exclude the current user and any logged in user who does not have an assigned character.

Additionally, this commit
- adds original macro attribution to Vindico#9013
- adds semantic emphasis to message wrapper
- adjusts formatting in chat card, including blockquoting whisper to make it easier to distinguish from surrounding advisory text
- removes token and 'checked' references