Prevent selection of characters that have no Corruption
The current and maximum Corruption of available characters is shown in the Send Dark Whisper dialog. Characters with no Corruption points are listed but cannot be selected for whispering.

Additional localisation strings are added to support user notification when there are no eligible characters to whisper to. In this case, the Send Dark Whisper dialog is not shown.