Optionally filter Dark Whisper targets by online status
Dark Whispers can be sent to offline users, ready for them to pick up when they log back in. Prior to this only active users could be targeted. The module default is to show all party members (both offline and online) in the Send Dark Whispers dialog.

A new module setting has added to allow GMs to set their preferred default filter option for Dark Whispers. The main function is parameterised and the user default can be easily overridden within the macro.