Abort Dark Whispers if no whisper message or target is set
If no whisper message or character was selected, a chat message would previously be sent with opening and closing text, but no ask. Now a tongue-in-cheek admonition is thrown up to the GM only if they run the macro. No chat message is generated or sent.

The validation is done after the dialog is closed, so it is possible for a GM to lose a whisper text if they complete that first and do not choose a target.

Additionally, some user facing content is changed:
- clearer advice about macro parameter options to override default character target list
- more flavourful dialog box prompts
- clearer user settings options and hint text for default character choice