Add example Dark Whispers
Each time the Dark Whispers dialog is fired, the table is rolled and a random result description is shown.

I originally had this shown as a placeholder but have chosen to add it as initial text so it can be copied or edited, as well as easily deleted. If user feedback warrants it, an option can be given in the future to set it as a placeholder instead.

This is the first GM Toolkit table. It's fairly meagre for now with barely a handful of results, but serves to feed sample Dark Whispers as a starting point and inspiration for the GM.

The table is a standard WFRP4e table for Foundry so can also be called directly from chat using the command `/table darkwhispers`.