Add support for players to confirm Dark Whispers response
A chat listener has been implemented and response buttons added to the whisper message that players get when they are sent a Dark Whisper. They can use these to confirm acceptance or refusal of the offer. A confirmation message is then sent to the chat log as a record of the character choice, including the original offer statement.

Mechanically, heeding a Dark Whisper reduces a character's Corruption by 1. This is not automated for now for technical and narrative reasons:
- There may be a delay between receiving an offer and fulfilling the terms. In these cases, removing the Corruption point would be premature. Instead the GM is able to make the call about when they consider the Whisper adhered to.
- Each whisper offer only removes a single Corruption. This would require some kind of lock on a chat button to track and avoid taking up the offer multiple times.
- If the chat card is Revealed to Everyone, any character would be able to accept the offer and take the Corruption perk.

Tracking that a character is committed to a Dark Whisper is entirely manual for now, beyond the chat log record. Going forward, this could be handled through some character sheet integration such as adding a Biography note, Personal Ambition or Extended Test.

Additional adjustments to support the Dark Whisper listener and improve UX include
- refactoring the message structure and translation keys to simplify and use string variables
- addition of a 'sendmode', to allow for whispers and broadcasts. Only private whisper messages are currently supported.
- addition of tooltip to explain why some characters cannot be selected.