Fix chat and notification messages
Changes have been made to translation keys and message construction that correct button placement and overal formatting in the Dark Whisper chat message that is sent to a target.

Notification messages and triggers have been improved throughout, particulary when attempting to accept or reject a deal if a GM (which is not permitted).

TODO: this could be improved by explicitly checking that the response button is being pressed by the targeted character. THis might involve traversing the path to  in the parent message.

TODO: Going one step further, changing the sequence of when the message is constructed in the macro (dark-whispers.js) could improve personalisation by incorporating character name (and id) within the message or metadata.